requirements.txt
cryptography==41.0.3
Pillow==9.4.0
pyzbar==0.1.9
qrcode==7.4.2
